Major code reorganization.

+++ video/__init__.py
@@ -16,13 +16,15 @@ Raspberry Pi.
 #   ~25kb reserved
 #   0% CPU usage
 
-import time
-import asyncio
-import subprocess
 import os
 import errno
+import asyncio
+import subprocess
+from time import time
 
-import tornado.web
+from tornado.ioloop import PeriodicCallback
+
+from printers import PrinterHandler
 
 try:
     import aiofiles
@@ -39,7 +41,7 @@ streams = {}
 stream_terminator = None
 
 
-async def start_streaming(app, path, name):
+async def start_streaming(printer, path):
     """
     Starts the streaming service for the given printer. This function is
     asynchronous and must be used with await since it doesn't complete until
@@ -47,37 +49,30 @@ async def start_streaming(app, path, name):
     """
 
     # Ensure path exists
-    if have_aiofiles:
-        await os.makedirs(path, exist_ok=True)
-    else:
-        os.makedirs(path, exist_ok=True)
+    if have_aiofiles: await os.makedirs(path, exist_ok=True)
+    else: os.makedirs(path, exist_ok=True)
 
-    # Get the video's url
-    if name not in app.config: raise tornado.web.HTTPError(404)
-    config = app.config[name]
-    if 'video' not in config: raise tornado.web.HTTPError(404)
-    url = config['video']
-    m3u8 = name + '.m3u8'
+    # Get the video's files
+    m3u8 = printer.name + '.m3u8'
     m3u8_full = os.path.join(path, m3u8)
 
     # Remove evidence of previous streaming
     try:
-        if have_aiofiles:
-            await os.remove(m3u8_full)
-        else:
-            os.remove(m3u8_full)
+        if have_aiofiles: await os.remove(m3u8_full)
+        else: os.remove(m3u8_full)
     except OSError as ex:
-        if ex.errno != errno.ENOENT:
-            raise
+        if ex.errno != errno.ENOENT: raise
 
-    print('Starting stream for '+name+'...')
+    print('Starting stream for '+printer.name+'...')
+    print(printer.get_video_url())
     # FFMPEG: https://www.ffmpeg.org/ffmpeg-formats.html#hls-2
     proc = subprocess.Popen((
-        'ffmpeg', '-hide_banner', '-nostats', '-loglevel', 'warning', '-i', url,
+        'ffmpeg', '-hide_banner', '-nostats', '-loglevel', 'error',
+        '-i', printer.get_video_url(),
         '-c:v', 'h264', '-profile:v', 'high', '-level', '4.1',
-        '-an', '-flags', '+cgop', '-g', '30',
-        '-hls_time', '2', '-hls_list_size', '3', '-hls_flags', 'delete_segments',
-        '-f', 'hls', m3u8
+        '-an', '-flags', '+cgop', '-g', '30', '-pix_fmt', 'yuv420p',
+        '-hls_time', '2', '-hls_list_size', '3',
+        '-hls_flags', 'delete_segments', '-f', 'hls', m3u8
     ), cwd=path)
 
     # Wait for the streaming to begin
@@ -88,9 +83,9 @@ async def start_streaming(app, path, name):
     return proc
 
 
-def terminate_streams(stale_secs=None):
+def terminate_video_streams(stale_secs=None):
     """Terminate all (stale) streams."""
-    stale = None if stale_secs is None else time.time() - stale_secs
+    stale = None if stale_secs is None else time() - stale_secs
     for name, info in streams.copy().items():
         if stale is None or info[1] < stale:
             print("Stopping stream for "+name+"...")
@@ -98,12 +93,12 @@ def terminate_streams(stale_secs=None):
             del streams[name]
 
 
-class VideoStaticFileHandler(tornado.web.StaticFileHandler):  # pylint: disable=abstract-method
+class VideoStaticFileHandler(PrinterHandler):  # pylint: disable=abstract-method
     def initialize(self, **kwargs):
         super().initialize(self.get_config('tmp'), **kwargs)
 
     def get_config(self, name):
-        val = self.application.config.get('VIDEO', name, fallback=None)
+        val = self.settings['config'].get('VIDEO', name, fallback=None)
         return DEFAULT_CONF[name] if val is None else val
 
 
@@ -111,26 +106,25 @@ class VideoHandler(VideoStaticFileHandler):  # pylint: disable=abstract-method
     """Handles *.m3u8 links which start the streaming service."""
 
     async def get(self, name, include_body=True):  # pylint: disable=arguments-differ
-
         # Make sure the terminator has started
         global stream_terminator
         if stream_terminator is None:
             keep_alive = int(self.get_config('keep-alive'))
-            stream_terminator = tornado.ioloop.PeriodicCallback(
-                lambda: terminate_streams(keep_alive*2), keep_alive*1000)
+            stream_terminator = PeriodicCallback(
+                lambda: terminate_video_streams(keep_alive*2), keep_alive*1000)
             stream_terminator.start()
         
         if name not in streams:
             # Streaming not currently running, start it
-            streams[name] = [None, time.time()]
-            proc = await start_streaming(self.application, self.root, name)
-            streams[name] = [proc, time.time()]
+            streams[name] = [None, time()]
+            proc = await start_streaming(self.get_printer(name), self.root)
+            streams[name] = [proc, time()]
         else:
             while streams[name] is None:
                 # Stream is being started right now, wait a little bit
                 await asyncio.sleep(0.001)
             
             # Stream is started, update last time accessed
-            streams[name][1] = time.time()
+            streams[name][1] = time()
 
         await super().get(name+'.m3u8', include_body)
